Move SnapshotBuilder from __init__ to utils.py
