+++ tests/ert/unit_tests/gui/experiments/test_run_dialog.py
@@ -54,9 +54,9 @@ from ert.run_models.event import (
 )
 from ert.run_models.run_model import RunModel
 from ert.scheduler.job import Job
-from tests.ert import SnapshotBuilder
 from tests.ert.handle_run_path_dialog import handle_run_path_dialog
 from tests.ert.ui_tests.gui.conftest import wait_for_child
+from tests.ert.utils import SnapshotBuilder
 
 _original_run_ensemble_evaluator_async = RunModel.run_ensemble_evaluator_async
